feat(core): add syslog input
